23: mentor recommendations

+++ lesson_23/js/tasks/six.js
@@ -3,7 +3,7 @@ const link = document.querySelector('.link')
 if (link) {
   link.dataset.someValue = '100'
 
-  const someValue = parseFloat(link.dataset.someValue)
+  const someValue = parseInt(link.dataset.someValue)
 
   if (someValue < 200) link.style.color = 'red'
 }
